Bug 1454864 - Improve merging of non-syncable items in the bookmarks mirror. r=markh,tcsc
This commit changes `fetchLocalTree` to include all items in Places,
syncable and non-syncable. The four user content roots, and all nodes
that descend from them, are syncable. Everything else is not.

When we check for structure changes of a local or remote node on the
other side, we also check if the node is syncable on both sides. If the
node is not syncable, on either side, we flag it for deletion. If the
non-syncable item is a child of a syncable parent (for example, an
orphaned left pane query), we set the parent's merge state to "new",
just like for local deletions, which flags the parent for reupload on
the next sync.

Uploading tombstones for non-syncable items is handled in the next
commit.

MozReview-Commit-ID: B9HzycuneND

UltraBlame original commit: 6cc5d9e64f7b733e47771753ffbbb405b18a1e46