Bug 1457678: Make the font canvas code not mess with the rule tree. r=xidorn,hiro
Inserting a lot of rules in the root level is super-inefficient. This fixes it
by not doing it.

It gives the root rule node to the style, but that's fine, since it's useless.
All this code-path is already pretty messy.

MozReview-Commit-ID: GoGHI4YJbKr

UltraBlame original commit: 32192f022b8455c322d592e9785359a7c9496c6a