Bug 1457027 - Part 1 - Don't persist the last selected handler. r=jaws
We plan to remove XUL persistence eventually, and this is a special case that uses the feature with an unusual attribute. This change also prevents pitfalls with getElementsByAttribute when the richlistitem descendants are made non-anonymous.

Some unneeded attributes in both the handlers and the containers richlistbox items are also cleaned up.

MozReview-Commit-ID: C05ArQGZb95

UltraBlame original commit: 33e9a37ae5b4b924459862106dcfa79b709b378d