Bug 1459714 - Use a single transaction for all things associated with a new display list. r=sotaro
This ensures the commands to delete compositables are received by WR in the same
transaction as the updates to the display list that removes the references to
those compositables. Otherwise, WR can try to do a scene build on the intermediate
state which results in missing pipeline errors.

MozReview-Commit-ID: ByOHCFWSSjt

UltraBlame original commit: 7b3c982608df4fca6b359e5a849c15feaf4de334