Bug 799710 - Move installer certificate validation to its own thread. r=agashlin
We occasionally get reports of UI unresponsiveness immediately following the
download phase of the stub installer. The longest operation that runs on the
main thread during this phase is validating the code signature of the full
installer. This patch moves that work (which is done in a native NSIS plugin)
to a separate thread. Hopefully this helps resolve the hangs.

I've also converted the build files for the plugin from Visual C++ 6 to 2017,
just to avoid the inconvenience of needing to pull up VC6 to build it.

MozReview-Commit-ID: CKje2a8M62i

UltraBlame original commit: 3af4e6789a7a1fef92c9f837ce52495add2f4aea