Bug 1434872 - [wdclient] Store original capabilities as used for the driver. r=ato
To allow Session.start() to be called multiple times the originally
requested capabilities for the driver have to be stored. Currently
those are overwritten with the ones as returned by the driver.

That means that custom entries which are getting extracted by the
driver (eg preferences) to setup the browser profile, are lost and
tests themselves cannot internally end a session and start it again.

MozReview-Commit-ID: Fi0cf6MaOc6

UltraBlame original commit: 14c67a3e28d5f400b85d47763b13b949a8b04502