Bug 1454640 - [docs] Lazy load the package and version properties r=mshal
We no longer store the docs under a project name (since all the docs are now
built using the root conf.py). This mean the name and version are only used for
packaging and uploading, which typically is only used in CI.

This allows us to lazy load the package name and version, so we only read the
conf.py when we need to.

MozReview-Commit-ID: DV5Jxrbskoh

UltraBlame original commit: d03f75986f626938142bd5bde293773506a2fc14