Bug 1454640 - [docs] Use a single SphinxManager instance across all rebuilds r=mshal
In the mozbuild.sphinx extension, we create a new SphinxManager instance each
time. However this isn't ideal now that we can rebuild the docs within the same
interpreter using the livereload server.

This makes use of a singleton so that we can share state not only between
multiple invocations of sphinx-build, but also with the mach command. This will
be taken advantage of more heavily in future commits in this series.

MozReview-Commit-ID: 7ERYeN5BPeI

UltraBlame original commit: 50bb4134a2e45dd02d44a0425116d0ef8c613f14