Bug 1455401 - Generate baseline code coverage reports. r=jmaher,marco
The tests added in this patch will be used as a baseline for coverage that is common to all tests. They are added to each chunk being run in the test-coverage suite only if a test with an associated baseline type exists.

MozReview-Commit-ID: 1CrRZ1Ev2Mz

UltraBlame original commit: 3d404958db3ccc82394f686681f4f00104e07b8f