Bug 1459850 - Wasm baseline: BaseCompiler::load() for 8-bit loads on x86: remove use of scratch reg.  r=lth.
For x86(32), BaseCompiler::load() loads indirectly via the scratch reg (ebx)
when the data size is 8 bits and the eventual destination is esi or edi.
This is pointless because the actual load created by |masm.wasmLoad(*access,
srcAddr, out);| is a mov{s,z}bl instruction and so is unconstrained as to
the output register.  This patch removes the indirection.

UltraBlame original commit: 1d050e53f0d3a22b44cd00e21911229ecb2ad5e0