Bug 1458383 - Assume tabs delay playback start of media. r=baku
Regardless of whether autoplay is enabled or not, we want to delay playback
start in tabs until the tab has been in the foreground. So initialize
nsPIDOMWindowOuter::mMediaSuspend to SUSPENDED_BLOCK if the controlling
pref for the delay-playback-start feature is enabled.

MozReview-Commit-ID: 7A8U5922AW8

UltraBlame original commit: 74dc3b5bb9aa99a041a48a9a096e05168b294bb6