Bug 1460400 - Don't enable dynamic builtin events on registration. r=chutten
This makes sure dynamic builtin events follow the same semantics as static builtin events.
On registration of the event the category is stored, but not enabled.
For fully-dynamic events, e.g. those registered by addons, the category
is enabled immediately (and can't be disabled).

This removes now-unused type definitions and switches from a map to a
simple set to store the category names.
The value stored in the map previously was not used at all.
In theory the map was effectively immutable after initialization, but
the check was only forced in debug anyway.
Now the set is mutable, but is only mutated in exactly 2 places.

MozReview-Commit-ID: 8tLEVXzHuHw

UltraBlame original commit: 843b90da1fad40d2379d327a26288663acd05539