Bug 1375829 part 2.  The default binding toJSON should skip over attributes that are not exposed in the current global.  r=qdot
Without this, we will start including mozMemory in performance.toJSON() even if
the pref for it is not set, once 'object' becomes a JSON type.

This changes behavior in the following observable ways:

1) We stop exposing PerformanceResourceTiming's .serverTiming in the JSON
serialization in insecure contexts.

2) We stop exposing PerformanceTiming's timeToNonBlankPaint and
timeToDOMContentFlushed in the JSON serialization unless the relevant
preferences are turned on.

UltraBlame original commit: 5a9cdc8de26aa2e643a05b312f9016a047219106