Bug 1462431 - Expand image downscale-on-decode to perform best effort sizing. r=tnikkel
When an image must be upscaled in only one dimension, we can
downscale-on-decode to a closer size than the native size. For example,
if an image is 32000x100, and we want 200x200, we can choose to decode
at 200x100. This allows drawing to scale in only one dimension, and use
less memory to store the decoded data, all the while not affecting the
display quality.

UltraBlame original commit: 6c8bf287d25b8d1ebc0392812b9328d57744233c