Bug 1415868 - remove ACTION_TASK r=jonasfj,tomprince
For kind=hook, the spec doesn't include this value as it's untrustworthy.

For kind=task, it's still untrustworthy, but there is no privilege escalation
so that's not important. Still, it dramatically expands the size of the task
definition.

MozReview-Commit-ID: 6scQ2ZwxP10

UltraBlame original commit: e2931d9049750c5a9da8caca77d4da65d085ef7b