Bug 1462469 - Add a "Clear Cookies and Site Data" footer button to the identity popup. r=nhnt11
This implements a new button in the identity popup that allows users to quickly
remove cookies and site data from the sites they're visiting.

This uses the SiteDataManager behind the scenes and is similar to the changes
we made for PageInfo already.

There's a major drawback to this approach in that SiteDataManager needs to
refresh its entire data set everytime we want information about a single site
or want to remove anything (it's not trivial to get rid of that limitation while
dealing with all the quirks of our storage APIs). I will work around this by
implementing a way for SiteDataManager to incrementally update itself in bug 1460768.

MozReview-Commit-ID: Iy7ia0KllFq

UltraBlame original commit: 59118f646da40f20320aadf81f7e9cd4cd92d70f