Bug 1462789.  Fix readystate assertions when document.open() is used on an image document.  r=hsivonen
There are a few things going on here:

1)  Better tracking of whether we've done our synthetic document stuff so we
don't redo it when it's not needed.  We should only be doing it on initial
load.

2)  Better handling of our readystate transitions.  We're only responsible for
the transition to READYSTATE_INTERACTIVE if we're the ones setting up the DOM.
Otherwise whoever is doing that should do the readystate transition.

Both of these really use the same mechanism, which is why I ended up doing them
together.

UltraBlame original commit: 9f9763a01686650308cbdf27d740de7065c5b5e8