Bug 1462791 - Add helper function for VCS checkout on generic-worker; r=dustin
This functionality was implemented at least 3 times. Let's consolidate
it to a central function.

Returning multiple command strings is kind of funky. I preserved
existing behavior and mozharness jobs are the only ones printing the
forensic logging. We should probably move this logging into
robustcheckout so we don't need to involve taskgraph with this. But
that can be deferred to another day.

MozReview-Commit-ID: I2LglJvfI6

UltraBlame original commit: 143be456f42025327408b28871585ed0d7b53742