Bug 1462500: Improve our detection of XPI vs. directory installs. r=aswan
This adds a helper so that we consistently detect install type based on
extension the same way.

In the process, it also simplifies some of the code that tries to detect
modification times for unpacked directories, so that it only checks the
directory mtime. We don't support signatures for unpacked directories anymore,
so this has no purpose outside of developer builds with signing disabled.
Those developers can move to touching the install directory's mtime rather than
the manifest file's.

MozReview-Commit-ID: J5LwaXXqyZL

UltraBlame original commit: cef00f3131bec5de42ab725cceb7821eef50b0a5