Bug 1457661 - P2. Ensure we call NotifyDataEnded for local resource once size is known. r=bryce
A call to NotifyDataEnded is required even if the size was known when the resource was created. This ensures that the readyState is properly updated and that playback can immediately as no more data can be added once first loaded.

MozReview-Commit-ID: FaJMBxJ9NkM

UltraBlame original commit: 27562071aaaa1175d70a03bc2f9214fee7e22829