Bug 1456946 - Don't call AudioNode::Initialize on an ChannelSplitterNode during construction to avoid throwing. r=achronop
We initialize the channelCountMode and channelInterpretation (that cannot be
changed) appropriately in the C++ ctor, and the channelCount manually as well.

Then we check manually that the options for the channelCountMode, the
channelCount and channelInterpretation have not been passed, or are not
incorrect.

MozReview-Commit-ID: 4YdYfaGqKqn

UltraBlame original commit: 4ac439f97956a46b322b85423c516d2eb6f17b2c