Bug 1466136: Don't look at the rule type from value parsing. r=hiro
This would cause properties to change the value semantics between, e.g.,
keyframes and non-keyframes, which would be observable.

It happens not to be observable since the animation-* and transition-*
properties are not allowed in keyframes, nor have bits in `contain`, and none
of the two properties are allowed in page. But I think it's the right thing to
do.

This still causes a quirk like a property value in chrome / user origins being
potentially different if the value is specified via CSS var functions. But I
think that is fine.

MozReview-Commit-ID: GhoPt0I34oO

UltraBlame original commit: b1c44c3fe871fbf61be12d160ccca2c08b91751b