Bug 1467134.  Use Atomic<bool> for the staticpref version of layout.css.font-variations.enabled.  r=emilio
Atomic<bool> is implemented in terms of AtomicBase<uint32_t>, because that way
you don't need to depend on atomic 1-byte operations.  That means that the rust
bindgen sees it as a u32, not a bool.

It's a bit concerning that the rust code seems to be doing an unsynchronized
read here, but given this is a RelaxedAtomic, that's probably ok.

UltraBlame original commit: 6de156e08e8e77f514185468e9364ea112fb6d4b