Bug 1463843 - Save the initial commited value in the first update call after text property editor calls create. r=pbro
In TextPropertyEditor, we call create() follow by update(). In create() we are parsing the
CSS property to store as the last committed values and this is unnnecessary and expensive
since we also parse the CSS property afterwards in update().

This changes moves the storing of the committed value over to the initial update() call
after create().

UltraBlame original commit: 6ce4f1c08f25f525cd1ce4b84878b6ffe29eb152