Bug 1468022 - Avoid quitting Firefox when enabled via pref at runtime. r=whimboo
If for whatever reason Marionette fails to initialise we quit
Firefox so that the parent process of Firefox (typically geckodriver)
does not have to elapse its connection timeout.

This patch rectifies an oversight that we also have this behaviour
when the user flips the marionette.enabled preference.  If the user
is in control of the process we should not forcefully make Firefox quit.

MozReview-Commit-ID: LVqAeIp56MG

UltraBlame original commit: 1e37948aa30cc61be6b5ea884234de2d3af9cd3f