Bug 1470020 - Part 4 - Don't handle Tracking Protection UI on file: and about: URIs. r=florian
... because these URIs are incompatible with TP.

We now show it on moz-extension: instead, which was forgotten previously.

This should probably have been in its own separate bug, but the changes in bug 1470020
surfaced this issue by throwing a lot of console errors when the baseURI was accessed,
so I didn't want to defer the fix.

MozReview-Commit-ID: 8KNV0oabv7Y

UltraBlame original commit: 6e0fc38940b484f165c4e76551d1bc91ded92ab7