Bug 1454973 - Part 2.Add the animation inspector test of RTL environment. r=daisuke
The patch of part 1 changed the logic of scrubber and tick label/line.
So this patch will add the related tests of this

Furthermoroe, If clicking the edge of scrubber area, the test might fail in RTL
environment due to a calculation of click position. This patch will add the 1px
margin-inline-start in RTL environment in order to prevent this case.

MozReview-Commit-ID: 4XcEBq80yk2

UltraBlame original commit: 7f7a190b3d0de55a0e231d9a32b82e2ce0643697